fuzz完善

+++ rosen/test/render_service/render_service/fuzztest/rsrenderserviceconnection005_fuzzer/rsrenderserviceconnection005_fuzzer.cpp
@@ -63,7 +63,7 @@ const uint8_t DO_REGISTER_FRAME_RATE_LINKER_EXPECTED_FPS_CALLBACK = 14;
 const uint8_t DO_REFRESH_RATE_MODE_CHANGE_CALLBACK = 15;
 const uint8_t DO_GET_REFRESH_INFO_TO_SP = 16;
 const uint8_t TARGET_SIZE = 17;
-const uint8_t WAIT_TASK_RUN_TIME_NS = 10000;
+const uint32_t WAIT_TASK_RUN_TIME_NS = 10000;
 
 sptr<RSIRenderServiceConnection> CONN = nullptr;
 const uint8_t* DATA = nullptr;
